# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RiverKin API",
    "description": "RiverKin backend \u2014 IEEE OneAquaHealth Global Hackathon 2026 (Track 5).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/crews": {
      "post": {
        "tags": [
          "crews"
        ],
        "summary": "Create Crew",
        "operationId": "create_crew_api_v1_crews_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrewCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crews/{crew_id}/members": {
      "post": {
        "tags": [
          "crews"
        ],
        "summary": "Add Member",
        "operationId": "add_member_api_v1_crews__crew_id__members_post",
        "parameters": [
          {
            "name": "crew_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Crew Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crews/{crew_id}/adoptions": {
      "post": {
        "tags": [
          "crews"
        ],
        "summary": "Adopt Site",
        "operationId": "adopt_site_api_v1_crews__crew_id__adoptions_post",
        "parameters": [
          {
            "name": "crew_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Crew Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdoptionIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Adopt Site Api V1 Crews  Crew Id  Adoptions Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crews/{crew_id}/checkins": {
      "post": {
        "tags": [
          "crews"
        ],
        "summary": "Checkin",
        "operationId": "checkin_api_v1_crews__crew_id__checkins_post",
        "parameters": [
          {
            "name": "crew_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Crew Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Checkin Api V1 Crews  Crew Id  Checkins Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crews/{crew_id}": {
      "get": {
        "tags": [
          "crews"
        ],
        "summary": "Get Crew",
        "operationId": "get_crew_api_v1_crews__crew_id__get",
        "parameters": [
          {
            "name": "crew_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Crew Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrewView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdoptedSiteOut": {
        "properties": {
          "site_code": {
            "type": "string",
            "title": "Site Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "days_unseen": {
            "type": "integer",
            "title": "Days Unseen"
          },
          "attention": {
            "type": "string",
            "title": "Attention"
          },
          "fresh": {
            "type": "boolean",
            "title": "Fresh"
          }
        },
        "type": "object",
        "required": [
          "site_code",
          "name",
          "days_unseen",
          "attention",
          "fresh"
        ],
        "title": "AdoptedSiteOut"
      },
      "AdoptionIn": {
        "properties": {
          "site_code": {
            "type": "string",
            "title": "Site Code"
          }
        },
        "type": "object",
        "required": [
          "site_code"
        ],
        "title": "AdoptionIn"
      },
...
      "CrewCreated": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CrewCreated"
      },
      "CrewIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "is_minor_crew": {
            "type": "boolean",
            "title": "Is Minor Crew",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CrewIn"
      },
      "CrewView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "is_minor_crew": {
            "type": "boolean",
            "title": "Is Minor Crew"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/MemberOut"
            },
            "type": "array",
            "title": "Members"
          },
          "adopted": {
            "items": {
              "$ref": "#/components/schemas/AdoptedSiteOut"
            },
            "type": "array",
            "title": "Adopted"
          },
          "coverage_pct": {
            "type": "integer",
            "title": "Coverage Pct"
          },
          "streak_windows": {
            "type": "integer",
            "title": "Streak Windows"
          },
          "checkin_active": {
            "type": "boolean",
            "title": "Checkin Active"
          },
          "simulated": {
            "type": "boolean",
            "title": "Simulated",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "city",
          "is_minor_crew",
          "members",
          "adopted",
          "coverage_pct",
          "streak_windows",
          "checkin_active"
        ],
        "title": "CrewView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MemberIn": {
        "properties": {
          "handle": {
            "type": "string",
            "title": "Handle"
          },
          "role_this_week": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role This Week"
          }
        },
        "type": "object",
        "required": [
          "handle"
        ],
        "title": "MemberIn"
      },
      "MemberOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "handle": {
            "type": "string",
            "title": "Handle"
          },
          "role_this_week": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role This Week"
          }
        },
        "type": "object",
        "required": [
          "id",
          "handle",
          "role_this_week"
        ],
        "title": "MemberOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}